Refactor into a more backend-agnostic API + clean-up.
